Memberlist for World of Warcraft Homepage
here is my code for a membership list for a website. Except for a few
little things this works quite well, but not all chars are read from the
guild. I have a diverenz of 17 chars.
cronfrage.php ---> reads the data from wow from
konfig.php ---> variable for the page
mitglieder.php ---> page to be opened with the table
mitglieder.css ---> I think explains itself